+++ appwrite_console/services/functions.py
@@ -263,7 +263,7 @@ class Functions(Service):
         Parameters
         ----------
         type : Optional[str]
-            Specification type to list. Can be one of: runtimes, builds.
+            Specification type to list. Can be one of: runtimes, builds. Defaults to runtimes.
         
         Returns
         -------